Create 2S4_TyulpanV3.py
Let's refactor the code by giving user a menu selection for ammo_type where they only select a number corresponding to the list so 1 = high-explosive, 2 = incendiary, and so on. Next we will modify the code to ask the user for the 53-F-864 fuze settings via list so 1 = delayed-action. When complete suggest a file name for the refactored code.

Let's refactor the code as per your requirements. We'll provide the user with a menu selection for ammo_type and fuze_setting by letting them select a number corresponding to the options.

Here's the modified code:

In this modified version of the code, the user is provided with a menu for ammo types and fuze settings by presenting the options and letting them select the number corresponding to their choice. The ammo type and fuze settings are then used for further calculations. Note that in this version, the fuze settings are not affecting the simulation, but it could be integrated into more complex modeling or simulation.