Create physics_charge_script.py
Let's refactor mortar_simulation.py by incorporating a physics model of the mortar firing. 

For example: We want to calculate the propellant charge necessary to propel a 240 mm caliber projectile with a length of 1635 mm and a mass of 134 kg to a distance of 9 kilometers.

To calculate the propellant charge required to propel a projectile, we need to determine the muzzle velocity required to reach the desired distance. Then, using the muzzle velocity and mass of the projectile, we can estimate the required energy. Finally, we'll determine the propellant charge needed to provide that energy. Generate all necessary equations to model this functionality and suggest a file name upon completion.

To calculate the propellant charge needed to propel a projectile to a certain distance, let's follow these steps:

Calculate Muzzle Velocity: Use kinematic equations to determine the muzzle velocity required to reach the desired distance. The range R of a projectile launched with a muzzle velocity v at an angle θ to the horizontal is given by:

R = (v² * sin(2θ)) / g

Here, g is the acceleration due to gravity (9.81 m/s²). We can use this equation to solve for the muzzle velocity v.

Calculate Required Energy: Next, calculate the energy required to achieve that muzzle velocity. The kinetic energy KE of an object with mass m and velocity v is given by:

KE = (1/2) * m * v²

Estimate Propellant Charge: Finally, estimate the propellant charge needed to provide that kinetic energy. The energy provided by the propellant is approximately proportional to its mass. Therefore, if we know the energy per unit mass of the propellant E_propellant, we can estimate the mass of the propellant m_propellant required as:

m_propellant = KE / E_propellant

Now let's integrate this physics model into the code. Note that E_propellant is an assumed value for the sake of this simulation. It is something that should be known from experimental data or from the manufacturer.

I have removed the plotting part for clarity and focused on the calculations related to propellant charge. Note that the energy per unit mass of the propellant (E_PROP) is an assumed value for this example and should be replaced with a more accurate value if available.

As for the file name, I would suggest naming it propellant_charge_calculator.py. This name reflects the new focus of the script, which is to calculate the propellant charge necessary for a given projectile and range.