Create real_time_mortar_simulation.py
To achieve real-time plotting of the trajectory, you can use FuncAnimation from the matplotlib.animation module. This allows you to create an animation by repeatedly calling a function (in this case, updating the position of the projectile).

Here is the refactored version of mortar_simulation.py with the added feature of real-time plotting:

This code will create an animation of the mortar shell's trajectory as if it is flying through the air in real-time.